Change tools links
