+++ pages/accountLists/[accountListId]/tools/fix/emailAddresses/[[...contactId]].page.test.tsx
@@ -2,16 +2,18 @@ import { useRouter } from 'next/router';
 import { ThemeProvider } from '@mui/material/styles';
 import { render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import { ErgonoMockShape } from 'graphql-ergonomock';
 import { getSession } from 'next-auth/react';
 import { SnackbarProvider } from 'notistack';
 import { I18nextProvider } from 'react-i18next';
+import { VirtuosoMockContext } from 'react-virtuoso';
 import TestRouter from '__tests__/util/TestRouter';
 import { GqlMockedProvider } from '__tests__/util/graphqlMocking';
-import { GetPersonDuplicatesQuery } from 'src/components/Tool/MergePeople/GetPersonDuplicates.generated';
-import { getPersonDuplicatesMocks } from 'src/components/Tool/MergePeople/PersonDuplicatesMock';
+import { mockInvalidEmailAddressesResponse } from 'src/components/Tool/FixEmailAddresses/FixEmailAddressesMocks';
+import { InvalidAddressesQuery } from 'src/components/Tool/FixMailingAddresses/GetInvalidAddresses.generated';
 import i18n from 'src/lib/i18n';
 import theme from 'src/theme';
-import MergePeoplePage from './[[...contactId]].page';
+import FixEmailAddressesPage from './[[...contactId]].page';
 
 jest.mock('next-auth/react');
 jest.mock('next/router', () => ({
@@ -33,6 +35,7 @@ jest.mock('notistack', () => ({
 
 const pushFn = jest.fn();
 const accountListId = 'account-list-1';
+const contactId = 'contactId';
 const session = {
   expires: '2021-10-28T14:48:20.897Z',
   user: {
@@ -42,25 +45,39 @@ const session = {
     token: 'superLongJwtString',
   },
 };
-const Components = () => (
+const Components = ({
+  mockNodes = mockInvalidEmailAddressesResponse,
+}: {
+  mockNodes?: ErgonoMockShape[];
+}) => (
   <ThemeProvider theme={theme}>
     <TestRouter>
-      <GqlMockedProvider<{
-        GetPersonDuplicates: GetPersonDuplicatesQuery;
-      }>
-        mocks={getPersonDuplicatesMocks}
+      <VirtuosoMockContext.Provider
+        value={{ viewportHeight: 1000, itemHeight: 100 }}
       >
-        <I18nextProvider i18n={i18n}>
-          <SnackbarProvider>
-            <MergePeoplePage />
-          </SnackbarProvider>
-        </I18nextProvider>
-      </GqlMockedProvider>
+        <GqlMockedProvider<{
+          InvalidAddresses: InvalidAddressesQuery;
+        }>
+          mocks={{
+            GetInvalidEmailAddresses: {
+              people: {
+                nodes: mockNodes,
+              },
+            },
+          }}
+        >
+          <I18nextProvider i18n={i18n}>
+            <SnackbarProvider>
+              <FixEmailAddressesPage />
+            </SnackbarProvider>
+          </I18nextProvider>
+        </GqlMockedProvider>
+      </VirtuosoMockContext.Provider>
     </TestRouter>
   </ThemeProvider>
 );
 
-describe('MergePeoplePage', () => {
+describe('FixEmailAddressesPage', () => {
   beforeEach(() => {
     (getSession as jest.Mock).mockResolvedValue(session);
     (useRouter as jest.Mock).mockReturnValue({
@@ -78,14 +95,14 @@ describe('MergePeoplePage', () => {
       expect(queryByTestId('loading')).not.toBeInTheDocument(),
     );
 
-    const contactName = await findByText('John Doe');
+    const contactName = await findByText('Test Contact');
 
     expect(contactName).toBeInTheDocument();
     userEvent.click(contactName);
 
     await waitFor(() => {
       expect(pushFn).toHaveBeenCalledWith(
-        `/accountLists/${accountListId}/tools/mergePeople/${'contact-1'}`,
+        `/accountLists/${accountListId}/tools/fix/emailAddresses/${contactId}`,
       );
     });
   });
