+++ pages/accountLists/[accountListId]/tools/fix/emailAddresses/[[...contactId]].page.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { loadSession } from 'pages/api/utils/pagePropsHelpers';
-import FixMailingAddresses from 'src/components/Tool/FixMailingAddresses/FixMailingAddresses';
-import { ToolsWrapper } from '../ToolsWrapper';
-import { SetContactFocus, useToolsHelper } from '../useToolsHelper';
+import { FixEmailAddresses } from 'src/components/Tool/FixEmailAddresses/FixEmailAddresses';
+import { ToolsWrapper } from '../../ToolsWrapper';
+import { SetContactFocus, useToolsHelper } from '../../useToolsHelper';
 
-const FixMailingAddressesPage: React.FC = () => {
+const FixEmailAddressesPage: React.FC = () => {
   const { t } = useTranslation();
   const { accountListId, handleSelectContact } = useToolsHelper();
-  const pageUrl = 'tools/fixMailingAddresses';
+  const pageUrl = 'tools/fix/emailAddresses';
 
   const setContactFocus: SetContactFocus = (contactId) => {
     handleSelectContact(pageUrl, contactId);
@@ -16,11 +16,11 @@ const FixMailingAddressesPage: React.FC = () => {
 
   return (
     <ToolsWrapper
-      pageTitle={t('Fix Mailing Addresses')}
+      pageTitle={t('Fix Email Addresses')}
       pageUrl={pageUrl}
-      selectedMenuId="fixMailingAddresses"
+      selectedMenuId="fixEmailAddresses"
     >
-      <FixMailingAddresses
+      <FixEmailAddresses
         accountListId={accountListId || ''}
         setContactFocus={setContactFocus}
       />
@@ -30,4 +30,4 @@ const FixMailingAddressesPage: React.FC = () => {
 
 export const getServerSideProps = loadSession;
 
-export default FixMailingAddressesPage;
+export default FixEmailAddressesPage;
