+++ pages/accountLists/[accountListId]/tools/fix/mailingAddresses/[[...contactId]].page.test.tsx
@@ -2,18 +2,17 @@ import { useRouter } from 'next/router';
 import { ThemeProvider } from '@mui/material/styles';
 import { render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { ErgonoMockShape } from 'graphql-ergonomock';
+import { ApolloErgonoMockMap } from 'graphql-ergonomock';
 import { getSession } from 'next-auth/react';
 import { SnackbarProvider } from 'notistack';
 import { I18nextProvider } from 'react-i18next';
-import { VirtuosoMockContext } from 'react-virtuoso';
 import TestRouter from '__tests__/util/TestRouter';
 import { GqlMockedProvider } from '__tests__/util/graphqlMocking';
-import { mockInvalidEmailAddressesResponse } from 'src/components/Tool/FixEmailAddresses/FixEmailAddressesMocks';
+import { mockInvalidAddressesResponse } from 'src/components/Tool/FixMailingAddresses/FixMailingAddressesMock';
 import { InvalidAddressesQuery } from 'src/components/Tool/FixMailingAddresses/GetInvalidAddresses.generated';
 import i18n from 'src/lib/i18n';
 import theme from 'src/theme';
-import FixEmailAddressesPage from './[[...contactId]].page';
+import FixMailingAddressesPage from './[[...contactId]].page';
 
 jest.mock('next-auth/react');
 jest.mock('next/router', () => ({
@@ -45,39 +44,25 @@ const session = {
     token: 'superLongJwtString',
   },
 };
-const Components = ({
-  mockNodes = mockInvalidEmailAddressesResponse,
-}: {
-  mockNodes?: ErgonoMockShape[];
-}) => (
+const Components = ({ mocks }: { mocks: ApolloErgonoMockMap }) => (
   <ThemeProvider theme={theme}>
     <TestRouter>
-      <VirtuosoMockContext.Provider
-        value={{ viewportHeight: 1000, itemHeight: 100 }}
+      <GqlMockedProvider<{
+        InvalidAddresses: InvalidAddressesQuery;
+      }>
+        mocks={mocks}
       >
-        <GqlMockedProvider<{
-          InvalidAddresses: InvalidAddressesQuery;
-        }>
-          mocks={{
-            GetInvalidEmailAddresses: {
-              people: {
-                nodes: mockNodes,
-              },
-            },
-          }}
-        >
-          <I18nextProvider i18n={i18n}>
-            <SnackbarProvider>
-              <FixEmailAddressesPage />
-            </SnackbarProvider>
-          </I18nextProvider>
-        </GqlMockedProvider>
-      </VirtuosoMockContext.Provider>
+        <I18nextProvider i18n={i18n}>
+          <SnackbarProvider>
+            <FixMailingAddressesPage />
+          </SnackbarProvider>
+        </I18nextProvider>
+      </GqlMockedProvider>
     </TestRouter>
   </ThemeProvider>
 );
 
-describe('FixEmailAddressesPage', () => {
+describe('FixMailingAddressesPage', () => {
   beforeEach(() => {
     (getSession as jest.Mock).mockResolvedValue(session);
     (useRouter as jest.Mock).mockReturnValue({
@@ -90,19 +75,27 @@ describe('FixEmailAddressesPage', () => {
   });
 
   it('should open up contact details', async () => {
-    const { findByText, queryByTestId } = render(<Components />);
+    const { getByText, queryByTestId } = render(
+      <Components
+        mocks={{
+          InvalidAddresses: {
+            ...mockInvalidAddressesResponse.InvalidAddresses,
+          },
+        }}
+      />,
+    );
     await waitFor(() =>
       expect(queryByTestId('loading')).not.toBeInTheDocument(),
     );
 
-    const contactName = await findByText('Test Contact');
+    const contactName = getByText('Baggins, Frodo');
 
     expect(contactName).toBeInTheDocument();
     userEvent.click(contactName);
 
     await waitFor(() => {
       expect(pushFn).toHaveBeenCalledWith(
-        `/accountLists/${accountListId}/tools/fixEmailAddresses/${contactId}`,
+        `/accountLists/${accountListId}/tools/fix/mailingAddresses/${contactId}`,
       );
     });
   });
