+++ pages/accountLists/[accountListId]/tools/fix/phoneNumbers/[[...contactId]].page.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { loadSession } from 'pages/api/utils/pagePropsHelpers';
 import FixPhoneNumbers from 'src/components/Tool/FixPhoneNumbers/FixPhoneNumbers';
-import { ToolsWrapper } from '../ToolsWrapper';
-import { SetContactFocus, useToolsHelper } from '../useToolsHelper';
+import { ToolsWrapper } from '../../ToolsWrapper';
+import { SetContactFocus, useToolsHelper } from '../../useToolsHelper';
 
 const FixPhoneNumbersPage: React.FC = () => {
   const { t } = useTranslation();
   const { accountListId, handleSelectContact } = useToolsHelper();
-  const pageUrl = 'tools/fixPhoneNumbers';
+  const pageUrl = 'tools/fix/phoneNumbers';
 
   const setContactFocus: SetContactFocus = (contactId) => {
     handleSelectContact(pageUrl, contactId);
