+++ pages/accountLists/[accountListId]/tools/merge/contacts/[[...contactId]].page.test.tsx
@@ -2,17 +2,16 @@ import { useRouter } from 'next/router';
 import { ThemeProvider } from '@mui/material/styles';
 import { render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { ApolloErgonoMockMap } from 'graphql-ergonomock';
 import { getSession } from 'next-auth/react';
 import { SnackbarProvider } from 'notistack';
 import { I18nextProvider } from 'react-i18next';
 import TestRouter from '__tests__/util/TestRouter';
 import { GqlMockedProvider } from '__tests__/util/graphqlMocking';
-import { mockInvalidNewslettersResponse } from 'src/components/Tool/FixSendNewsletter/FixSendNewsletterMock';
-import { InvalidNewsletterQuery } from 'src/components/Tool/FixSendNewsletter/InvalidNewsletter.generated';
+import { GetContactDuplicatesQuery } from 'src/components/Tool/MergeContacts/GetContactDuplicates.generated';
+import { getContactDuplicatesMocks } from 'src/components/Tool/MergeContacts/MergeContactsMock';
 import i18n from 'src/lib/i18n';
 import theme from 'src/theme';
-import FixSendNewsletterPage from './[[...contactId]].page';
+import MergeContactsPage from './[[...contactId]].page';
 
 jest.mock('next-auth/react');
 jest.mock('next/router', () => ({
@@ -31,6 +30,7 @@ jest.mock('notistack', () => ({
     };
   },
 }));
+
 const pushFn = jest.fn();
 const accountListId = 'account-list-1';
 const session = {
@@ -45,22 +45,22 @@ const session = {
 const Components = () => (
   <ThemeProvider theme={theme}>
     <TestRouter>
-      <I18nextProvider i18n={i18n}>
-        <SnackbarProvider>
-          <GqlMockedProvider<{
-            InvalidNewsletter: InvalidNewsletterQuery;
-          }>
-            mocks={mockInvalidNewslettersResponse as ApolloErgonoMockMap}
-          >
-            <FixSendNewsletterPage />
-          </GqlMockedProvider>
-        </SnackbarProvider>
-      </I18nextProvider>
+      <GqlMockedProvider<{
+        GetContactDuplicates: GetContactDuplicatesQuery;
+      }>
+        mocks={getContactDuplicatesMocks}
+      >
+        <I18nextProvider i18n={i18n}>
+          <SnackbarProvider>
+            <MergeContactsPage />
+          </SnackbarProvider>
+        </I18nextProvider>
+      </GqlMockedProvider>
     </TestRouter>
   </ThemeProvider>
 );
 
-describe('FixSendNewsletterPage', () => {
+describe('MergeContactsPage', () => {
   beforeEach(() => {
     (getSession as jest.Mock).mockResolvedValue(session);
     (useRouter as jest.Mock).mockReturnValue({
@@ -75,16 +75,17 @@ describe('FixSendNewsletterPage', () => {
   it('should open up contact details', async () => {
     const { findByText, queryByTestId } = render(<Components />);
     await waitFor(() =>
-      expect(queryByTestId('LoadingSpinner')).not.toBeInTheDocument(),
+      expect(queryByTestId('loading')).not.toBeInTheDocument(),
     );
-    const contactName = await findByText('Baggins, Frodo');
+
+    const contactName = await findByText('Doe, John');
 
     expect(contactName).toBeInTheDocument();
     userEvent.click(contactName);
 
     await waitFor(() => {
       expect(pushFn).toHaveBeenCalledWith(
-        `/accountLists/${accountListId}/tools/fixSendNewsletter/${'contactId1'}`,
+        `/accountLists/${accountListId}/tools/merge/contacts/${'contact-1'}`,
       );
     });
   });
