+++ pages/accountLists/[accountListId]/tools/merge/contacts/[[...contactId]].page.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { loadSession } from 'pages/api/utils/pagePropsHelpers';
-import MergePeople from 'src/components/Tool/MergePeople/MergePeople';
-import { ToolsWrapper } from '../ToolsWrapper';
-import { SetContactFocus, useToolsHelper } from '../useToolsHelper';
+import MergeContacts from 'src/components/Tool/MergeContacts/MergeContacts';
+import { ToolsWrapper } from '../../ToolsWrapper';
+import { SetContactFocus, useToolsHelper } from '../../useToolsHelper';
 
-const MergePeoplePage: React.FC = () => {
+const MergeContactsPage: React.FC = () => {
   const { t } = useTranslation();
   const { accountListId, handleSelectContact } = useToolsHelper();
-  const pageUrl = 'tools/mergePeople';
+  const pageUrl = 'tools/merge/contacts';
 
   const setContactFocus: SetContactFocus = (contactId) => {
     handleSelectContact(pageUrl, contactId);
@@ -16,9 +16,9 @@ const MergePeoplePage: React.FC = () => {
 
   return (
     <ToolsWrapper
-      pageTitle={t('Merge People')}
+      pageTitle={t('Merge Contacts')}
       pageUrl={pageUrl}
-      selectedMenuId="mergePeople"
+      selectedMenuId="mergeContacts"
       styles={
         <style>{`
           div.MuiBox-root {
@@ -31,7 +31,7 @@ const MergePeoplePage: React.FC = () => {
     `}</style>
       }
     >
-      <MergePeople
+      <MergeContacts
         accountListId={accountListId || ''}
         setContactFocus={setContactFocus}
       />
@@ -41,4 +41,4 @@ const MergePeoplePage: React.FC = () => {
 
 export const getServerSideProps = loadSession;
 
-export default MergePeoplePage;
+export default MergeContactsPage;
