+++ src/components/Tool/Home/ToolList.ts
@@ -48,26 +48,26 @@ export const ToolsList: ToolsGroup[] = [
         desc: 'Set the correct contacts commitment info for each contact',
 
         icon: mdiCurrencyUsd,
-        id: 'fixCommitmentInfo',
+        id: 'fix/commitmentInfo',
       },
       {
         tool: 'Fix Mailing Addresses',
         desc: 'Set the correct primary mailing address for each contact',
 
         icon: mdiMap,
-        id: 'fixMailingAddresses',
+        id: 'fix/mailingAddresses',
       },
       {
         tool: 'Fix Send Newsletter',
         desc: 'Set the correct newsletter state for each contact',
         icon: mdiNewspaperVariantOutline,
-        id: 'fixSendNewsletter',
+        id: 'fix/sendNewsletter',
       },
       {
         tool: 'Merge Contacts',
         desc: 'Review and merge duplicate contacts',
         icon: mdiHome,
-        id: 'mergeContacts',
+        id: 'merge/contacts',
       },
     ],
   },
@@ -79,19 +79,19 @@ export const ToolsList: ToolsGroup[] = [
         tool: 'Fix Email Addresses',
         desc: 'Set the correct primary email address for each person',
         icon: mdiEmail,
-        id: 'fixEmailAddresses',
+        id: 'fix/emailAddresses',
       },
       {
         tool: 'Fix Phone Numbers',
         desc: 'Set the correct primary phone number for each person',
         icon: mdiPhone,
-        id: 'fixPhoneNumbers',
+        id: 'fix/phoneNumbers',
       },
       {
         tool: 'Merge People',
         desc: 'Review and merge duplicate people',
         icon: mdiAccountGroup,
-        id: 'mergePeople',
+        id: 'merge/people',
       },
     ],
   },
